warning code simplified
